% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ckiener}
\alias{ckiener}
\alias{ckiener1}
\alias{ckiener2}
\alias{ckiener3}
\alias{ckiener4}
\title{Fat Tail Corrective Function for Kiener Distributions}
\usage{
ckiener(p, k, e)

ckiener1(p, k)

ckiener2(p, a, w)

ckiener3(p, k, d)

ckiener4(p, k, e)
}
\arguments{
\item{p}{vector of probabilities.}

\item{k}{numeric. parameter k used in model k1, k3 and k4.}

\item{e}{numeric. parameter d used in model k4.}

\item{a}{numeric. parameter a used in model k2.}

\item{w}{numeric. parameter w used in model k2.}

\item{d}{numeric. parameter d used in model k3.}
}
\description{
Quantile functions of Kiener distributions K1, K2, K3 and K4 can be expressed
as a combination of logit function and fat tail corrective function.
}
\details{
Quantile functions of Kiener distributions K1, K2, K3 and K4 can be expressed
as a combination of logit function and fat tail corrective function.
}
\seealso{
\code{\link{logit}}, \code{\link{qlogis}}, \code{\link{qkiener1}},
              \code{\link{qkiener2}}, \code{\link{qkiener3}}, \code{\link{qkiener4}}.
}

