% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/k_estimation.R
\name{estimkappa6}
\alias{estimkappa6}
\title{Local Function estimkappa6}
\usage{
estimkappa6(lg, lp, dx1p, dx1g, dxg, dxp, maxk = 10)
}
\arguments{
\item{lg}{logit of probability close to the median.}

\item{lp}{logit of extreme probability.}

\item{dx1p}{numeric. Delta of quantile m - x(1-p).}

\item{dx1g}{numeric. Delta of quantile m - x(1-pg).}

\item{dxg}{numeric. Delta of quantile x(pg) - m.}

\item{dxp}{numeric. Delta of quantile x(p) - m.}

\item{maxk}{numeric. Maximum allowed value for k (kappa).}
}
\description{
Internal function to estimate parameter k (kappa).
}
\details{
Internal function to estimate parameter k (kappa). Called by \code{estimkappa11}.
}

