% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_data.R
\docType{data}
\name{pprobs0}
\alias{pprobs0}
\alias{pprobs1}
\alias{pprobs2}
\alias{pprobs3}
\alias{pprobs4}
\alias{pprobs5}
\alias{pprobs6}
\alias{pprobs7}
\alias{pprobs8}
\alias{pprobs9}
\title{Several Vectors of Probabilities}
\format{An object of class \code{numeric} of length 9.}
\usage{
pprobs0

pprobs1

pprobs2

pprobs3

pprobs4

pprobs5

pprobs6

pprobs7

pprobs8

pprobs9
}
\description{
Several vectors of probabilities used in FatTailsR. 
Remark: pprobs5 <- sort(c(pprobs2, pprobs3, pprobs4)).

pprobs0 <- c(0.01, 0.05, 0.10, 0.25, 0.50, 0.75, 0.90, 0.95, 0.99)

pprobs1 <- c(0.01, 0.05, 0.95, 0.99)

pprobs2 <- c(0.01, 0.025, 0.05, 0.95, 0.975, 0.99)

pprobs3 <- c(0.001, 0.0025, 0.005, 0.995, 0.9975, 0.999)

pprobs4 <- c(0.0001, 0.00025, 0.0005, 0.9995, 0.99975, 0.9999)

pprobs5 <- c(0.0001, 0.00025, 0.0005, 0.001, 0.0025, 0.005, 0.01, 0.025, 0.05, 
0.95, 0.975, 0.99, 0.995, 0.9975, 0.999, 0.9995, 0.99975, 0.9999)

pprobs6 <- c(0.0001, 0.0005, 0.001, 0.005, 0.01, 0.05, 0.50, 
0.95, 0.99, 0.995, 0.999, 0.9995, 0.9999) 

pprobs7 <- c(0.01, 0.025, 0.05, 
0.10, 0.17, 0.25, 0.33, 0.41, 0.50, 0.59, 0.67, 0.75, 0.83, 0.90, 
0.95, 0.975, 0.99) 

pprobs8 <- c(0.001, 0.0025, 0.005, 0.01, 0.025, 0.05, 
0.10, 0.17, 0.25, 0.33, 0.41, 0.50, 0.59, 0.67, 0.75, 0.83, 0.90, 
0.95, 0.975, 0.99, 0.995, 0.9975, 0.999) 

pprobs9 <- c(0.0001, 0.00025, 0.0005, 0.001, 0.0025, 0.005, 0.01, 0.025, 0.05, 
0.10, 0.17, 0.25, 0.33, 0.41, 0.50, 0.59, 0.67, 0.75, 0.83, 0.90, 
0.95, 0.975, 0.99, 0.995, 0.9975, 0.999, 0.9995, 0.99975, 0.9999)
}
\seealso{
The conversion function \code{\link{getnamesk}}
}
\keyword{datasets}
