% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_data.R
\docType{data}
\name{tData}
\alias{tData}
\title{Datasets dfData, mData, tData, xData, zData, extractData : tData}
\description{
A list of datasets in data.frame, matrix, timeSeries, xts and zoo formats. 
This is the timeSeries format. 
Visit \code{\link{extractData}} for more information.
}
\keyword{datasets}
