% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_data.R
\docType{data}
\name{xData}
\alias{xData}
\title{Datasets dfData, mData, tData, xData, zData, extractData : xData}
\description{
A list of datasets in data.frame, matrix, timeSeries, xts and zoo formats. 
This is the xts format. 
Visit \code{\link{extractData}} for more information.
}
\keyword{datasets}
