% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/k_estimation.R
\name{checkquantiles}
\alias{checkquantiles}
\title{Local Function Checkquantiles}
\usage{
checkquantiles(x)
}
\arguments{
\item{x}{vector of quantiles}
}
\description{
Internal function to verify that quantiles are all different from
each other and correctly ordered.
}

