% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/b_data.R
\docType{data}
\name{dfData}
\alias{dfData}
\title{Datasets dfData, tData, xData, zData, extractData : dfData}
\description{
A list of datasets in data.frame, timeSeries, xts and zoo formats.
This is the data.frame format.
Visit \code{\link{extractData}} for more information.
}
\keyword{datasets}

