% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/b_data.R
\docType{data}
\name{xData}
\alias{xData}
\title{Datasets dfData, tData, xData, zData, extractData : xData}
\description{
A list of datasets in data.frame, timeSeries, xts and zoo formats.
This is the xts format.
Visit \code{\link{extractData}} for more information.
}
\keyword{datasets}

