% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/z_data.R
\docType{data}
\name{xData}
\alias{xData}
\title{xData}
\description{
A list of datasets in data.frame, timeSeries, xts and zoo formats.
This is the xts format.
Visit \code{\link{dfData}} for more information.
}
\keyword{datasets}

