% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_data.R
\docType{data}
\name{zData}
\alias{zData}
\title{Datasets dfData, mData, tData, xData, zData, extractData : zData}
\description{
A list of datasets in data.frame, matrix, timeSeries, xts and zoo formats. 
This is the zoo format. 
Visit \code{\link{extractData}} for more information.
}
\keyword{datasets}
